backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-jose
python-multipart
python-multipart
httpx
playwright
pytest
pytest-asyncio
httpx
